rename consts to events
