+++ integrations/jenkins/webhook/webhook_processors/build_webhook_processor.py
@@ -1,4 +1,4 @@
-from webhook.consts import BUILD_DELETE_EVENTS, BUILD_UPSERT_EVENTS
+from webhook.events import BUILD_DELETE_EVENTS, BUILD_UPSERT_EVENTS
 from webhook.webhook_processors.jenkins_abstract_webhook_processor import (
     _JenkinsAbstractWebhookProcessor,
 )
